Add menu to main(), renaming
Add menu to main() method. Rename some classes and add a new package
