+++ java/src/lab5/task_5_1/scanner/InputScanner.java
@@ -1,10 +1,10 @@
-package lab5.task_5_1.menu;
+package lab5.task_5_1.scanner;
 
 import java.util.Scanner;
 
 public class InputScanner {
 	
-	public static String scan() {
+	public static String scanInput() {
 		Scanner scanner = new Scanner(System.in);
 		String input = scanner.nextLine();
 		scanner.close();
